Add Experiment controls to MatchedSet pages
Waiting for a question in the ticket to be answered before issuing a PR.
